(PC-13743) feat(storybook): change StyledInputContainer > InputContainer and InputContainer > ContainerWithMaxWidth
