+++ src/ui/components/inputs/DropDown/DropDown.web.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { v4 as uuidv4 } from 'uuid'
 
 import { InputLabel } from 'ui/components/InputLabel.web'
-import { InputContainer } from 'ui/components/inputs/InputContainer'
+import { ContainerWithMaxWidth } from 'ui/components/inputs/ContainerWithMaxWidth'
 import { ArrowDown } from 'ui/svg/icons/ArrowDown'
 import { getSpacingString, Spacer } from 'ui/theme'
 
@@ -36,7 +36,7 @@ export function DropDown({
   const dropDownInputID = uuidv4()
 
   return (
-    <InputContainer>
+    <ContainerWithMaxWidth>
       <InputLabel htmlFor={dropDownInputID}>{label}</InputLabel>
       <Spacer.Column numberOfSpaces={2} />
       <SelectContainer>
@@ -62,7 +62,7 @@ export function DropDown({
           <ArrowDown />
         </IconContainer>
       </SelectContainer>
-    </InputContainer>
+    </ContainerWithMaxWidth>
   )
 }
